license: switch to ISC license
There are several BSD licenses: 4-clause (non free), 3-clause (used by
this project), 2-clause, 1-clause. Moreover, the 3-clause versions has
a non-endorsement clause where a name should be put and it is always a
bit confused on what name should be put here.

This commit switches to a simpler license, the ISC license. Only one
version of this license exists and no templating is needed for it.